Starting cplusplus

+++ graphutil.c
@@ -8,13 +8,13 @@
  * @param rows
  * @return
  */
-short* buildIndicesForGLLINEs(int vcount, int *rows, int rowCount, int *size, int *loggedSize){
+int* buildIndicesForGLLINEs(int vcount, int *rows, int rowCount, int *size, int *loggedSize){
 	int i, j, colCount=0, k;
-	short *indices;
-	short *tmp;
+	int *indices;
+	int *tmp;
 	
 	*loggedSize = 100;
-	indices = (short*)malloc(sizeof(short) * (*loggedSize));
+	indices = (int*)malloc(sizeof(int) * (*loggedSize));
 	k = 0;
 	*size = 0;
 	for(i=0; i<rowCount; i++) {
@@ -22,14 +22,14 @@ short* buildIndicesForGLLINEs(int vcount, int *rows, int rowCount, int *size, in
 		for(j = 0; j<colCount; j++) {
 			if(*size >= (*loggedSize)) {
 				(*loggedSize) += 100;
-				tmp = (short*)realloc(indices, sizeof(short) * (*loggedSize));
+				tmp = (int*)realloc(indices, sizeof(int) * (*loggedSize));
 				indices = tmp;
 			}
 			indices[(*size)++] = k;
 			if(j>0 && j<colCount-1) {
 				if(*size >= (*loggedSize)) {
 					(*loggedSize) += 100;
-					tmp = (short*)realloc(indices, sizeof(short) * (*loggedSize));
+					tmp = (int*)realloc(indices, sizeof(int) * (*loggedSize));
 					indices = tmp;
 				}
 				indices[(*size)++] = k;
@@ -43,14 +43,14 @@ short* buildIndicesForGLLINEs(int vcount, int *rows, int rowCount, int *size, in
 		for(i=0; i<rowCount; i++) {
 			if(*size >= (*loggedSize)) {
 				(*loggedSize) += 100;
-				tmp = (short*)realloc(indices, sizeof(short) * (*loggedSize));
+				tmp = (int*)realloc(indices, sizeof(int) * (*loggedSize));
 				indices = tmp;
 			}
 			indices[(*size)++] = j + colCount*i;
 			if(i> 0 && i<rowCount-1) {
 				if(*size >= (*loggedSize)) {
 					(*loggedSize) += 100;
-					tmp = (short*)realloc(indices, sizeof(short) * (*loggedSize));
+					tmp = (int*)realloc(indices, sizeof(int) * (*loggedSize));
 					indices = tmp;
 				}
 				indices[(*size)++] = j + colCount*i;
