+++ graphutil.h
@@ -1,6 +1,6 @@
 #ifndef __GRAPHUTIL_H_
 #define __GRAPHUTIL_H_
 
-short* buildIndicesForGLLINEs(int vcount, int *rows, int rowCount, int *size, int *loggedSize);
+int* buildIndicesForGLLINEs(int vcount, int *rows, int rowCount, int *size, int *loggedSize);
 
 #endif
