+++ linux/nmath.h
@@ -12,7 +12,7 @@ extern "C" {
 typedef struct tagDParam{
 	NMAST *t;
 	char variables[4];
-	short error;
+	int error;
 	DATA_TYPE_FP *values;
 	DATA_TYPE_FP retv;
 	NMAST *returnValue;
@@ -20,7 +20,7 @@ typedef struct tagDParam{
 
 void initFunct(Function *);
 void toString(const NMAST *t, char *str, int *curpos, int len);
-void resetFunction(Function *f, const char *str, const char *vars, int varCount, short *error);
+void resetFunction(Function *f, const char *str, const char *vars, int varCount, int *error);
 DATA_TYPE_FP calc(Function *f, DATA_TYPE_FP *values, int numOfValue, int *);
 int reduce(Function *f, int *);
 #ifdef WINDOWS
